bump sparktool spark version to 3.4.0
