--- Utils/spark-tools/spark-v2.3/src/test/scala/org/apache/spark/SparkContextWithFailureSaveSuite.scala
@@ -1,46 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation
- *
- * All rights reserved.
- *
- * MIT License
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
- * documentation files (the "Software"), to deal in the Software without restriction, including without limitation
- * the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and
- * to permit persons to whom the Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of
- * the Software.
- *
- * THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO
- * THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
- * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package org.apache.spark
-
-import org.apache.spark.scheduler.DAGWithFailureSaveScheduler
-import org.scalatest.concurrent.Eventually
-import org.scalatest.MustMatchers._
-
-class SparkContextWithFailureSaveSuite extends SparkFunSuite with LocalSparkContext with Eventually {
-  def mySC: SparkContextWithFailureSave = sc.asInstanceOf[SparkContextWithFailureSave]
-
-  test("SparkContextWithFailureSave should use paired DAGWithFailureSaveScheduler scheduler") {
-    val conf = new SparkConf().setAppName("test").setMaster("local")
-    sc = new SparkContextWithFailureSave(conf)
-
-    sc.dagScheduler mustBe a[DAGWithFailureSaveScheduler]
-  }
-
-  test("SparkContextWithFailureSave should save the broadcast values") {
-    val conf = new SparkConf().setAppName("test").setMaster("local")
-    sc = new SparkContextWithFailureSave(conf)
-
-    val bc = sc.broadcast("hello broadcast world!")
-    mySC.bcIdMap must contain value bc
-  }
-}
