+++ Utils/spark-tools/spark-v3.4/src/main/scala/com/microsoft/azuretools/SparkToolsAnchor.scala
