+++ Utils/spark-tools/spark-v3.4/src/main/scala/org/apache/spark/failure/FailureTask.scala
